Extract Micro-Practice routes into routes/microPractice.ts
Moves all 15 /api/practice/* and /api/students/:id/micro-practice/*
(plus weak-competencies, which shares a helper) routes out of index.ts's
startServer() closure, matching the routes/students.ts pattern.
calculateNextScheduleState moves into practiceScheduleService.ts,
exported for direct unit testing, alongside the schedule logic it
belongs with.

Also includes a Tailwind conditional-class indentation fix in
Layout.tsx left over from the earlier upstream merge.
